requirements.txt
Django==1.10
djangorestframework==3.5.3
